Use double instead of float in -seekToPosition:

+++ Player/SFBAudioPlayer.mm
@@ -385,7 +385,7 @@ namespace {
 	return [_playerNode seekToTime:timeInSeconds];
 }
 
-- (BOOL)seekToPosition:(float)position
+- (BOOL)seekToPosition:(double)position
 {
 	return [_playerNode seekToPosition:position];
 }
